+++ R/utils.R
@@ -1,8 +1,8 @@
-#' Setup the ASU example
+#' Create the ASU example
 #' 
-#' This code sets up the Arizona example IPU problem and avoids repeating this
-#' code in multiple places throughout the package (vignettes/tests).
-#' @return When run, it creates four variables in the calling environment:
+#' Sets up the Arizona example IPU problem and is used in multiple places
+#' throughout the package (vignettes/tests).
+#' @return A list of four variables:
 #'   hh_seed, hh_targets, per_seed, and per_targets. These can be used directly
 #'   by \code{\link{ipu}}.
 #' @export