+++ inst/doc/common_ipf_problems.R
@@ -1,16 +1,16 @@
-## ---- include = FALSE----------------------------------------------------
+## ---- include = FALSE---------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
 
-## ----setup---------------------------------------------------------------
+## ----setup--------------------------------------------------------------------
 suppressPackageStartupMessages({
   library(ipfr)
   library(dplyr)
 })
 
-## ---- warning=TRUE-------------------------------------------------------
+## ---- warning=TRUE------------------------------------------------------------
 hh_seed <- tibble(
   geo_region = 1,
   id = c(1:8),
@@ -33,10 +33,10 @@ hh_targets$hhveh <- tibble(
 
 result <- ipu(hh_seed, hh_targets, max_iterations = 30, verbose = TRUE)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 result$primary_comp
 
-## ----balance example inputs----------------------------------------------
+## ----balance example inputs---------------------------------------------------
 result <- setup_arizona()
 hh_seed <- result$hh_seed
 hh_targets <- result$hh_targets
@@ -46,7 +46,7 @@ per_targets <- result$per_targets
 avg_hh_weight <- (rowSums(hh_targets$hhtype) - 1) / nrow(hh_seed)
 avg_per_weight <- (rowSums(per_targets$pertype) - 1) / nrow(per_seed)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 new_per_targets <- per_targets
 new_per_targets$pertype <- per_targets$pertype %>%
   mutate_at(
@@ -56,10 +56,10 @@ new_per_targets$pertype <- per_targets$pertype %>%
 
 result <- ipu(hh_seed, hh_targets, per_seed, new_per_targets, max_iterations = 30)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 result$weight_dist
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 result_80 <- ipu(
   hh_seed, hh_targets, per_seed, new_per_targets,
   max_iterations = 30,
@@ -70,7 +70,7 @@ result_80$weight_dist
 result_80$primary_comp
 result_80$secondary_comp
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 result_20 <- ipu(
   hh_seed, hh_targets, per_seed, new_per_targets,
   max_iterations = 30,
@@ -81,7 +81,7 @@ result_20$weight_dist
 result_20$primary_comp
 result_20$secondary_comp
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 hh_seed <- tibble(
   id = c(1, 2, 3, 4),
   siz = c(1, 2, 2, 1),
@@ -99,13 +99,13 @@ hh_targets$siz <- tibble(
 result <- ipu(hh_seed, hh_targets, max_iterations = 10,
               max_ratio = 1.2, min_ratio = .8)
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 result$weight_dist
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 result$primary_comp
 
-## ------------------------------------------------------------------------
+## -----------------------------------------------------------------------------
 hh_targets <- list()
 hh_targets$siz <- tibble(
   geo_cluster = c(1, 2),