+++ NEWS.md
@@ -1,3 +1,8 @@
-# ipfr 1.0.0 (2019-08-05)
+# ipfr 1.0.1 (2019-12-12)
+
+  * Fixed bug when marginal target only had 1 category (#2)
+  * Minor update to keep package working with latest r-devel
+
+# ipfr 1.0.0 (2019-08-21)
 
   * First release of the package.